Add TestSource support to dynamic containers and tests
This commit introduces the support to add an instance of `TestSource`
to a dynamic container or test by adding static factory methods that
take a test source as a parameter.

Prior to this commit the source for dynamic tests was always a
`MethodSource` pointing to the `@TestFactory` annotated method.
Now, for example, it is possible to refer to external files, that
are the underlying sources of generated containers and tests.

Addresses part of: #1178